refactor(calculator): 사용자 입력을 long 범위까지 커버할 수 있도록 변경
문제 요구 상에서는 단순히 '구분자와 양수로 구성된 문자열'이라고 명시되어 있다. 즉, 숫자의 상한 범위가 지정되어 있지 않다.

이에 대해서 "애매하다면 최대한 넓게 포용할 수 있도록 한다"라는 판단을 하겠다. 즉, 기존의 int 타입을 long으로 변경한다.

long보다 더 큰 BigInteger도 존재하지만, 기본형이 아니기에 거기까지 생각하는 것은 과하다고 판단한다.

기본형을 넘어선 범위까지 커버해야 된다면 숫자 범위 등으로 힌트를 주지 않았을까 생각된다.

+++ src/main/java/calculator/StringCalculator.java
@@ -2,11 +2,11 @@ package calculator;
 
 public class StringCalculator {
 
-    public static int calculate(String userInput) {
-        int[] separated = StringSeparator.separate(userInput);
+    public static long calculate(String userInput) {
+        long[] separated = StringSeparator.separate(userInput);
 
-        int sum = 0;
-        for (int number : separated) {
+        long sum = 0;
+        for (long number : separated) {
             sum += number;
         }
         return sum;
