+++ src/test/java/calculator/StringCalculatorTest.java
@@ -15,7 +15,7 @@ public class StringCalculatorTest {
         String userInput = "1,2:3";
 
         // when
-        int result = StringCalculator.calculate(userInput);
+        long result = StringCalculator.calculate(userInput);
 
         // then
         assertThat(result).isEqualTo(6);
@@ -28,7 +28,7 @@ public class StringCalculatorTest {
         String userInput = "";
 
         // when
-        int result = StringCalculator.calculate(userInput);
+        long result = StringCalculator.calculate(userInput);
 
         // then
         assertThat(result).isEqualTo(0);
@@ -41,7 +41,7 @@ public class StringCalculatorTest {
         String userInput = "//;\\n1,2:3;4";
 
         // when
-        int result = StringCalculator.calculate(userInput);
+        long result = StringCalculator.calculate(userInput);
 
         // then
         assertThat(result).isEqualTo(10);
